requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.17.0
streamlit-chat==0.1.1
tiktoken==0.5.2
chromadb==0.4.18 